# api/openapi.yaml
openapi: "3.0.3"
info:
  title: Uploy API
  version: 0.2.0
  description: Uploy deployment platform API
paths:
...
  /api/deployments:
    post:
      operationId: createDeployment
      summary: Create a new deployment
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeployRequest"
      responses:
        "200":
          description: Deployment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeployResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Service not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: The server proxy must be upgraded before a domain deployment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/deployments/{id}/logs:
    get:
      operationId: getDeploymentLogs
      summary: Stream deployment logs via SSE
      description: |
        Returns a Server-Sent Events stream. OpenAPI does not natively model SSE,
        so this endpoint is documented for reference only.

        **SSE events (not modeled by this schema):**
        - Default message event: JSON-encoded LogEntry
        - `done` event: deployment final status string ("success" or "failed")
        - `stream-error` event: JSON object with "message" field
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: SSE log stream
          content:
            text/event-stream:
              schema:
                $ref: "#/components/schemas/LogEntry"
        "400":
          description: Invalid deployment ID format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Deployment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/projects:
    post:
      operationId: createProject
      summary: Create a new project
      security:
        - cookieAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: listProjects
      summary: List projects in workspace
      security:
        - cookieAuth: []
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/services/{id}/deployments:
    get:
      operationId: listServiceDeployments
      summary: List deployment history for a service
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of deployments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DeploymentResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Service not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
...
    DeployRequest:
      type: object
      required:
        - service_id
      properties:
        service_id:
          type: string
    DeployResponse:
      type: object
      required:
        - deployment_id
      properties:
        deployment_id:
          type: string
    DeploymentResponse:
      type: object
      required:
        - id
        - status
        - service_id
        - created_at
        - phase
        - is_active
        - is_draining
        - is_rolling
      properties:
        id:
          type: string
        status:
          type: string
          enum:
            - in_progress
            - success
            - failed
        service_id:
          type: string
        created_at:
          type: string
          format: date-time
        phase:
          type: string
        is_active:
          type: boolean
        is_draining:
          type: boolean
        is_rolling:
          type: boolean
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
        code:
          type: string
          description: >
            Machine-readable failure stage. Present on SSH probe failures so a
            client can offer the right remedy: the three stages need entirely
            different fixes, and the human message alone cannot be branched on.
          enum:
            - unreachable
            - key_rejected
            - key_invalid
            - session_failed
            - docker_missing
...
    LogEntry:
      type: object
      required:
        - id
        - order
        - created_at
        - output
        - type
        - phase
      properties:
        id:
          type: integer
          format: int64
        order:
          type: integer
        created_at:
          type: string
          format: date-time
        output:
          type: string
        type:
          type: string
          enum:
            - stdout
            - stderr
        phase:
          type: string
          description: Structured deploy phase identifier. Empty string for raw command output.
    ProjectResponse:
      type: object
      required:
        - id
        - name
        - workspace_id
        - created_at
        - updated_at
      properties:
        id:
          type: string
        name:
          type: string
        workspace_id:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateProjectRequest:
      type: object
      description: |
        Request body for creating a project. When `name` is omitted, blank, or
        whitespace-only, the backend generates a unique Railway-style
        `adjective-noun` name scoped to the workspace.
      properties:
        name:
          type: string
...